todo-v2/requirements.txt
a
asgiref
attrs
certifi
charset-normalizer
cheroot
CherryPy
decorator
distlib
dj-static
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
gunicorn
idna
install
jaraco.classes
jaraco.collections
jaraco.context
jaraco.functools
jaraco.text
jsonschema
more-itertools
of
Pillow
portend
PyJWT
pymongo
pyrsistent
python-decouple
pytz
requests
six
sqlparse
static3
tempora
to
urllib3
ws4py
zc.lockfile
